# Refactor : improved perf at launch, cleaned up some code