Hi,

This PR contains two commit.

* **[Refactor] Replaced try/catch(KeyNotFoundEception) with if (ContainsKey)**
  * Some `try/catch(KeyNotFoundEception)` block are replaced by `if (array.containsKey)/else` blocks. This will greatly improve perfs at launch.

* **[Refactor] Removed references to old TQ save path, trash and TQVaultMon**
  * With AE, all characters are "Immortal Throne by default". This mean it's possible to simplify some methods !
  * Removed references to [trash panel](https://blog-imgs-96-origin.fc2.com/k/e/i/keibro2/2016101002293376b.jpg). This feature was removed long ago. (The screenshot bring back some memories, for sure ! 😄)
  * Removed various unused pieces of code.

That's still a pleasure to contribute to TQVault, even if TQ is an old game, it's awesome to see that this project is still active !