@NorthFury 2018-11-11
In regards to removing the check for the existence of the Immortal Throne expansion, at first I thought that it may still be useful for people having old saving games. I stand corrected because those saves will be automatically converted the first time they are loaded in the most recent version of TQ.

I agree that relying on exceptions is far more costly in terms of performance. This is a nice fix.
I would have expected to be very rare that the keys are not found. Can you notice a significant improvement in startup time? Once you confirm this I'll approve and merge the PR.