@KenAlin 2018-11-12
Your concern about IT seems valid. I can't really say how TQVaultAE will react with old saves, but if TQ AE convert the saves at first launch, that's perfect.

About exceptions : in fact, TQVault read TQ Database *via* these exceptions. To give you a number : Visual Studio debugger counts more than 9000 `KeyNotFoundException` raised before the vault screen appears.

About performances, here are some quick (and maybe dirty) results : 
- With exceptions : 4s / 26s
- Without exceptions : 2s / 5s
(normal launch / launch via VS debugger)